separate fixes and building html doc from the main script.
We take the parts of the code that are responsible for building the html
documents into separate modules.
